+++ tabulous/_qt/_mainwindow/_command_palette.py
@@ -67,6 +67,7 @@ for cmd, desc in [
     (cmds.table.random, "Generate random values"),
     (cmds.table.undo_table, "Undo table"),
     (cmds.table.redo_table, "Redo table"),
+    (cmds.table.show_undo_stack_view, "Show undo stack view"),
 ]:
     table_group.register(_command_to_function(cmd), desc=desc)
 
@@ -115,33 +116,33 @@ for cmd, desc in [
 
 
 for cmd, desc in [
-    (cmds.table.copy_data_tab_separated, "Copy cells (tab separated)"),
-    (cmds.table.copy_data_with_header_tab_separated, "Copy cells with headers (tab separated)"),
-    (cmds.table.copy_data_comma_separated, "Copy cells (comma separated)"),
-    (cmds.table.copy_data_with_header_comma_separated, "Copy cells with headers (comma separated)"),
-    (cmds.table.copy_as_new_table, "Copy as a new table"),
-    (cmds.table.copy_as_new_spreadsheet, "Copy as a new spreadsheet"),
-    (cmds.table.copy_as_literal, "Copy as literal"),
-    (cmds.table.paste_data_tab_separated, "Paste from tab separated text"),
-    (cmds.table.paste_data_comma_separated, "Paste from comma separated text"),
-    (cmds.table.paste_data_from_numpy_string, "Paste from numpy-style text"),
-    (cmds.table.select_all, "Select all the cells"),
-    (cmds.table.cut_data, "Cut selected cells"),
-    (cmds.table.delete_values, "Delete selected cells"),
-    (cmds.table.add_highlight, "Add highlight to selected cells"),
-    (cmds.table.insert_row_above, "Insert row above current index"),
-    (cmds.table.insert_row_below, "Insert row below current index"),
-    (cmds.table.insert_column_left, "Insert column left of current index"),
-    (cmds.table.insert_column_right, "Insert column right of current index"),
-    (cmds.table.remove_this_row, "Remove current row"),
-    (cmds.table.remove_this_column, "Remove current column"),
-    (cmds.table.set_column_dtype, "Set column dtype"),
-    (cmds.table.set_foreground_colormap, "Set foreground colormap"),
-    (cmds.table.set_background_colormap, "Set background colormap"),
-    (cmds.table.reset_foreground_colormap, "Reset foreground colormap"),
-    (cmds.table.reset_background_colormap, "Reset background colormap"),
-    (cmds.table.set_text_formatter, "Set text formatter"),
-    (cmds.table.reset_text_formatter, "Reset text formatter"),
+    (cmds.selection.copy_data_tab_separated, "Copy cells (tab separated)"),
+    (cmds.selection.copy_data_with_header_tab_separated, "Copy cells with headers (tab separated)"),
+    (cmds.selection.copy_data_comma_separated, "Copy cells (comma separated)"),
+    (cmds.selection.copy_data_with_header_comma_separated, "Copy cells with headers (comma separated)"),
+    (cmds.selection.copy_as_new_table, "Copy as a new table"),
+    (cmds.selection.copy_as_new_spreadsheet, "Copy as a new spreadsheet"),
+    (cmds.selection.copy_as_literal, "Copy as literal"),
+    (cmds.selection.paste_data_tab_separated, "Paste from tab separated text"),
+    (cmds.selection.paste_data_comma_separated, "Paste from comma separated text"),
+    (cmds.selection.paste_data_from_numpy_string, "Paste from numpy-style text"),
+    (cmds.selection.select_all, "Select all the cells"),
+    (cmds.selection.cut_data, "Cut selected cells"),
+    (cmds.selection.delete_values, "Delete selected cells"),
+    (cmds.selection.add_highlight, "Add highlight to selected cells"),
+    (cmds.selection.insert_row_above, "Insert row above current index"),
+    (cmds.selection.insert_row_below, "Insert row below current index"),
+    (cmds.selection.insert_column_left, "Insert column left of current index"),
+    (cmds.selection.insert_column_right, "Insert column right of current index"),
+    (cmds.selection.remove_this_row, "Remove current row"),
+    (cmds.selection.remove_this_column, "Remove current column"),
+    (cmds.selection.set_column_dtype, "Set column dtype"),
+    (cmds.selection.set_foreground_colormap, "Set foreground colormap"),
+    (cmds.selection.set_background_colormap, "Set background colormap"),
+    (cmds.selection.reset_foreground_colormap, "Reset foreground colormap"),
+    (cmds.selection.reset_background_colormap, "Reset background colormap"),
+    (cmds.selection.set_text_formatter, "Set text formatter"),
+    (cmds.selection.reset_text_formatter, "Reset text formatter"),
 ]:
     selection_group.register(_command_to_function(cmd), desc=desc)
