+++ tabulous/commands/__init__.py
@@ -1,4 +1,4 @@
 from __future__ import annotations
 
 
-from . import io, plot, tab, table, view, analysis, window
+from . import io, plot, selection, tab, table, view, analysis, window
